poker_ocr2.py
from PIL import Image
import pytesseract
import re

# Note: This is a simplified script and will likely require significant adjustments
# based on the actual image quality, font styles, and layout variations.
# Image processing for this level of detail is complex and often requires
# more advanced techniques and libraries like OpenCV.
# This example uses PIL for basic image loading and pytesseract for OCR,
# which might not be accurate enough for all the data points.

# You might need to install these libraries:
# pip install Pillow pytesseract

# Also, you might need to configure tesseract-ocr on your system.

def extract_poker_data(image_path):
    """
    Attempts to extract relevant poker data from the given image path.
    This is a basic implementation and might not be highly accurate.
    """
    try:
        img = Image.open(image_path)
        text = pytesseract.image_to_string(img)
        print("Extracted Text:\n", text)

        data = {}

        # **Warning:** This is a very basic and unreliable way to extract data.
        # Real-world image analysis for structured data like this requires
        # more robust techniques (e.g., template matching, object detection,
        # analyzing specific image regions).

        # Attempt to find the blinds information
        blinds_match = re.search(r"HOLD'EM - (\d+)/(\d+)", text)
        if blinds_match:
            data["blinds"] = f"{blinds_match.group(1)}/{blinds_match.group(2)}"

        # Attempt to find the pot size
        pot_match = re.search(r"Pot\s*(\d+BB)", text)
        if pot_match:
            data["pot"] = int(pot_match.group(1).replace("BB", ""))

        # Attempt to find player information (very basic and likely inaccurate)
        player_matches = re.findall(r"(\w+)\s*(\d+\.?\d*BB)", text)
        players = []
        for match in player_matches:
            players.append({"username": match[0], "stack": float(match[1].replace("BB", ""))})
        data["players"] = players

        # Attempt to find the current player's cards (very basic and unreliable)
        card_match = re.search(r"(\d+[SHDC])\s*(\d+[SHDC])", text) # Example pattern
        if card_match:
            data["current_player_cards"] = [card_match.group(1), card_match.group(2)]

        # Attempt to find betting options (very basic and unreliable)
        if "Fold" in text:
            data["betting_options"] = {"fold": True}
            call_match = re.search(r"Call\s*(\d+BB)", text)
            if call_match:
                data["betting_options"]["call"] = int(call_match.group(1).replace("BB", ""))
            raise_match = re.search(r"Raise\s*(\d+BB)", text)
            if raise_match:
                data["betting_options"]["raise"] = int(raise_match.group(1).replace("BB", ""))

        return data

    except Exception as e:
        print(f"Error processing image: {e}")
        return None

def send_data(data):
    """
    This function simulates sending the extracted data.
    In a real application, this would involve sending the data to an API,
    storing it in a database, or performing other actions.
    """
    if data:
        print("Data successfully received by send_data function.")
        return True
    else:
        print("No data received by send_data function.")
        return False

if __name__ == "__main__":
    image_file = "/Users/trip/git/PokerJoker/test1.jpeg"  # Replace with the actual path to your image
    extracted_data = extract_poker_data(image_file)

    if extracted_data:
        print("\nExtracted Data:")
        print(extracted_data)
        send_data(extracted_data)
    else:
        print("\nFailed to extract data from the image.")