requirements.txt
opencv-python
pytesseract
requests
numpy
Pillow
pyautogui